move sync to message broker
