+++ dev/tests/api-functional/testsuite/Magento/CatalogStorefront/ProductVariantsTest.php
@@ -5,7 +5,7 @@
  */
 declare(strict_types=1);
 
-namespace Magento\CatalogMessageBroker\Model\MessageBus;
+namespace Magento\CatalogStorefront;
 
 use Magento\Catalog\Api\Data\ProductInterface;
 use Magento\Catalog\Api\ProductRepositoryInterface;
