+++ dev/tests/api-functional/testsuite/Magento/CatalogStorefront/ProductsTest.php
@@ -6,7 +6,7 @@
 
 declare(strict_types=1);
 
-namespace Magento\CatalogMessageBroker\Model\MessageBus;
+namespace Magento\CatalogStorefront;
 
 use Magento\Catalog\Api\Data\ProductInterface;
 use Magento\Catalog\Api\ProductRepositoryInterface;
